+++ app/routes/carts.js
@@ -3,154 +3,16 @@ const router = express.Router();
 
 const checkAuth = require('../middleware-auth/check-auth');
 
-const Cart = require('../models/carts');
+const CartsController = require('../controllers/carts');
 
-router.post('/', checkAuth, (req, res) => {
-    // Create a cart
-    const cart = new Cart({
-        productName: req.body.map(({ productName }) => productName),
-        // quantity: req.body.quantity,
-        totalAmount: req.body.totalAmount,
-        meta: { ...req.body.meta, created: new Date() }
-    });
-    console.log('carrrrrrrrr', req.body);
-    cart
-        .save()
-        .then(payload => {
-            console.log(payload);
-            res.status(201).json({
-                message: 'Added product cart successfully',
-                createdCart: {
-                    productName: payload.productName,
-                    quantity: payload.quantity,
-                    totalAmount: req.body.totalAmount,
-                    uuid: payload.uuid,
-                    request: {
-                        type: 'GET',
-                        url: 'http://localhost:3000/carts/' + payload.uuid
-                    }
-                }
-            });
-        })
-        .catch(err => {
-            console.log(err);
-            res.status(500).json({
-                error: err
-            })
-        });
-});
+router.post('/', checkAuth, CartsController.cartsPOST);
 
-router.get('/', checkAuth, (req, res) => {
-    Cart.find({'meta.active': { $gte: true }}, {__v: 0, _id: 0})
-    .exec()
-    .then(payload => {
-        const response = {
-            count: payload.length,
-            products: payload.map(payload => {
-                return {
-                    productName: req.body.productName,
-                    quantity: req.body.quantity,
-                    totalAmount: req.body.totalAmount,
-                    uuid: payload.uuid,
-                    request: {
-                        type: 'GET',
-                        url: 'http://localhost:3000/carts/' + payload.uuid
-                    }
-                }}),
-        };
-        console.log(payload)
-        // if (payload) {
-            res.status(200).json({
-                response
-        })
-        // } else {
-        //     res.status(404).json({
-        //         message: 'Not found'
-        //     })
-        // }
-    })
-    .catch(err => {
-        console.log(err);
-        res.status(500).json({
-            error: err
-        });
-    });
-});
+router.get('/', checkAuth, CartsController.cartsGetAll);
 
-router.get(`/:productUUID`, checkAuth, (req, res) => {
-    const uuid = req.params.productUUID;
-    Cart.findOne({ uuid, 'meta.active': { $gte: true } }, {__v: 0, _id: 0})
-    .exec()
-    .then(payload => {
-        console.log("From database", payload);
-        if (payload) {
-            res.status(200).json({
-                product: payload,
-                request: {
-                    type: 'GET',
-                    description: 'Get a Single product(item) from Cart',
-                    url: 'http://localhost:3000/carts/' + payload.uuid
-            }});
-        } else {
-            res.status(404).json({ message: 'Not found'});
-        }
-    })
-    .catch(err => {
-        console.log(err);
-        res.status(500).json({error: err});
-    });
-});
+router.get(`/:productUUID`, checkAuth, CartsController.cartsGetOne);
 
-router.put('/:productUUID', checkAuth, (req, res) => {
-    const uuid = req.params.productUUID;
-    const updateOps = {};
-    for (const ops of req.body){
-        updateOps[ops.propName] = ops.value;
-    }
-    Cart.updateOne({ uuid: uuid }, { $set: updateOps, 'meta.updated': Date.now()})
-        .exec()
-        .then(payload => {
-            res.status(200).json({
-                cart: payload,
-                message: 'Product in cart updated',
-                request: {
-                    type: 'PUT',
-                    description: 'Updates a Single product in cart',
-                    url: 'http://localhost:3000/carts/' + uuid
-        }})
-    })
-    .catch(err => {
-        res.status({
-            error: err
-        })
-    })
-});
+router.put('/:productUUID', checkAuth, CartsController.cartsUpdateOne);
 
-router.delete('/:productUUID', checkAuth, (req, res) => {
-    const uuid = req.params.productUUID;
-    Cart.deleteOne(
-        { uuid },
-        {
-            $set: {'meta.updated': new Date() }
-        }
-    )
-    .exec()
-    .then(payload => {
-            res.status(200).json({
-                product: payload,
-                message: 'Product deleted',
-                request: {
-                    type: 'DELETE',
-                    description: 'Deletes a single product by its uuid',
-                    url: 'http://localhost:3000/products/'
-        }})
-    })
-    .catch(err => {
-        console.log(err);
-        res.status(500).json({
-            error: err
-        })
-    });
-});
+router.delete('/:productUUID', checkAuth, CartsController.cartsDeleteOne);
 
 module.exports = router;
